Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int a=5;
-    int b=10;
-    int c=15;
+    float a=5;
+    float b=10;
+    float c=15;
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
